# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BookIt API
  version: 1.0.0
  description: REST API сервиса онлайн-записи BookIt.
servers:
  - url: http://localhost:3000/api
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/MessageResponse"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            token:
              type: string
          required:
            - user
            - token
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        service_id:
          type: integer
        booking_date:
          type: string
          example: "2026-06-15"
        booking_time:
          type: string
          example: "14:00"
        status:
          type: string
          enum: [active, cancelled, completed]
        created_at:
          type: string
          format: date-time
        service_name:
          type: string
        service_price:
          type: integer
        service_duration:
          type: integer
        service_desc:
          type: string
        user_name:
          type: string
...
    AdminStats:
      type: object
      properties:
        users:
          type: integer
        services:
          type: integer
        bookings:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            cancelled:
              type: integer
        revenue:
          type: integer
        popular_services:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
        active_bookings:
          type: integer
        total_bookings:
          type: integer
paths:
  /register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: newuser
                password:
                  type: string
                  example: pass1234
                timezone:
                  type: string
                  example: Europe/Moscow
      responses:
        "201":
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Логин занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      tags: [Auth]
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: demo
                password:
                  type: string
                  example: user123
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /me:
    get:
      tags: [Auth]
      summary: Текущий пользователь
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Данные текущего пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Требуется авторизация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Создать запись
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, booking_date, booking_time]
              properties:
                service_id:
                  type: integer
                booking_date:
                  type: string
                  example: "2026-06-15"
                booking_time:
                  type: string
                  example: "14:00"
      responses:
        "201":
          description: Запись создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: "#/components/schemas/Booking"
                  message:
                    type: string
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Требуется авторизация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Слот занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/my:
    get:
      tags: [Bookings]
      summary: Мои записи
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Список записей
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    delete:
      tags: [Bookings]
      summary: Отменить запись
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Запись отменена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Нельзя отменить чужую запись
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Запись не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/admin/all:
    get:
      tags: [Bookings]
      summary: Все записи
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Все записи
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
                  stats:
                    type: object
  /bookings/busy-times:
    get:
      tags: [Bookings]
      summary: Занятые слоты
      parameters:
        - in: query
          name: service_id
          required: true
          schema:
            type: integer
        - in: query
          name: date
          required: true
          schema:
            type: string
            example: "2026-06-15"
      responses:
        "200":
          description: Занятые и доступные слоты
          content:
            application/json:
              schema:
                type: object
                properties:
                  busy_times:
                    type: array
                    items:
                      type: string
                  slots:
                    type: array
                    items:
                      type: object
...
  /admin/stats:
    get:
      tags: [Admin]
      summary: Статистика системы
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Статистика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      tags: [Admin]
      summary: Пользователи с поиском и пагинацией
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
            enum: [all, user, admin]
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
...